ekho-backend/requirements.txt
# Core Framework
fastapi==0.121.0
uvicorn[standard]==0.32.1

# Google Cloud - Latest versions
google-cloud-aiplatform==1.90.0
google-cloud-storage==2.18.2
google-auth==2.35.0
google-generativeai

# Data Validation - Pydantic V2
pydantic==2.12.4
pydantic-settings==2.7.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.1
python-multipart==0.0.20
pillow==11.0.0

# MongoDB
motor

# Snowflake
snowflake-connector-python==4.0.0

# ElevenLabs
elevenlabs